+++ PyPong/setting.py
@@ -1,4 +1,5 @@
 import os
+import pygame
 
 display_width = 1080
 display_height = 720
@@ -6,11 +7,11 @@ display_height = 720
 
 BLACK = (0, 0, 0)
 WHITE = (255, 255, 255)
+BACKGROUND = '#190781'
 
 images_dict = {}
-
-for path, dir, file in os.walk('pypong_images'):
-    print(path, file)
+# получаем доступные изображения
+for path, dir, file in os.walk(r'./pypong_images'):
     for name_file in file:
-        images_dict[name_file] = os.path.join(path, name_file)
+        images_dict[name_file] = path + '/' + name_file
 print(images_dict)
